Cleanup tests and move common functions to utility file.
