+++ tests/mlmodel_openai/test_chat_completion_error_v1.py
@@ -14,11 +14,12 @@
 
 import openai
 import pytest
-from conftest import (
-    disabled_ai_monitoring_record_content_settings,
+from testing_support.ml_testing_utils import (
+    add_token_count_to_events,
     events_sans_content,
-    add_token_count_to_event,
     llm_token_count_callback,
+    disabled_ai_monitoring_record_content_settings,
+    set_trace_info,
 )
 from testing_support.fixtures import (
     dt_enabled,
@@ -336,7 +337,7 @@ def test_chat_completion_invalid_request_error_invalid_model(set_trace_info, syn
     rollup_metrics=[("Llm/completion/OpenAI/create", 1)],
     background_task=True,
 )
-@validate_custom_events(add_token_count_to_event(expected_events_on_invalid_model_error))
+@validate_custom_events(add_token_count_to_events(expected_events_on_invalid_model_error))
 @validate_custom_event_count(count=2)
 @background_task()
 def test_chat_completion_invalid_request_error_invalid_model_with_token_count(set_trace_info, sync_openai_client):
@@ -417,7 +418,7 @@ def test_chat_completion_invalid_request_error_invalid_model_async(loop, set_tra
     rollup_metrics=[("Llm/completion/OpenAI/create", 1)],
     background_task=True,
 )
-@validate_custom_events(add_token_count_to_event(expected_events_on_invalid_model_error))
+@validate_custom_events(add_token_count_to_events(expected_events_on_invalid_model_error))
 @validate_custom_event_count(count=2)
 @background_task()
 def test_chat_completion_invalid_request_error_invalid_model_with_token_count_async(
