+++ tests/mlmodel_openai/test_chat_completion_stream_v1.py
@@ -13,13 +13,14 @@
 # limitations under the License.
 
 import openai
-from conftest import (  # pylint: disable=E0611
-    add_token_count_to_event,
-    disabled_ai_monitoring_record_content_settings,
-    disabled_ai_monitoring_settings,
-    disabled_ai_monitoring_streaming_settings,
+from testing_support.ml_testing_utils import (
+    add_token_count_to_events,
     events_sans_content,
     llm_token_count_callback,
+    disabled_ai_monitoring_record_content_settings,
+    disabled_ai_monitoring_streaming_settings,
+    disabled_ai_monitoring_settings,
+    set_trace_info,
 )
 from testing_support.fixtures import (
     reset_core_stats_engine,
@@ -185,7 +186,7 @@ def test_openai_chat_completion_sync_in_txn_with_llm_metadata_no_content(set_tra
 
 @reset_core_stats_engine()
 @override_llm_token_callback_settings(llm_token_count_callback)
-@validate_custom_events(add_token_count_to_event(chat_completion_recorded_events))
+@validate_custom_events(add_token_count_to_events(chat_completion_recorded_events))
 # One summary event, one system message, one user message, and one response message from the assistant
 @validate_custom_event_count(count=4)
 @validate_transaction_metrics(
@@ -472,7 +473,7 @@ def test_openai_chat_completion_async_conversation_id_set_no_content(loop, set_t
 
 @reset_core_stats_engine()
 @override_llm_token_callback_settings(llm_token_count_callback)
-@validate_custom_events(add_token_count_to_event(chat_completion_recorded_events))
+@validate_custom_events(add_token_count_to_events(chat_completion_recorded_events))
 # One summary event, one system message, one user message, and one response message from the assistant
 # @validate_custom_event_count(count=4)
 @validate_transaction_metrics(
