+++ tests/mlmodel_openai/test_embeddings_error.py
@@ -14,11 +14,12 @@
 
 import openai
 import pytest
-from conftest import (
-    add_token_count_to_event,
-    disabled_ai_monitoring_record_content_settings,
+from testing_support.ml_testing_utils import (
+    add_token_count_to_events,
     events_sans_content,
     llm_token_count_callback,
+    disabled_ai_monitoring_record_content_settings,
+    set_trace_info,
 )
 from testing_support.fixtures import (
     dt_enabled,
@@ -182,7 +183,7 @@ invalid_model_events = [
     ],
     background_task=True,
 )
-@validate_custom_events(add_token_count_to_event(invalid_model_events))
+@validate_custom_events(add_token_count_to_events(invalid_model_events))
 @validate_custom_event_count(count=1)
 @background_task()
 def test_embeddings_invalid_request_error_invalid_model_with_token_count(set_trace_info):
@@ -450,7 +451,7 @@ def test_embeddings_invalid_request_error_no_model_async_no_content(loop, set_tr
     ],
     background_task=True,
 )
-@validate_custom_events(add_token_count_to_event(invalid_model_events))
+@validate_custom_events(add_token_count_to_events(invalid_model_events))
 @validate_custom_event_count(count=1)
 @background_task()
 def test_embeddings_invalid_request_error_invalid_model_with_token_count_async(set_trace_info, loop):
