+++ tests/mlmodel_openai/test_embeddings_v1.py
@@ -13,12 +13,13 @@
 # limitations under the License.
 
 import openai
-from conftest import (  # pylint: disable=E0611
-    add_token_count_to_event,
-    disabled_ai_monitoring_record_content_settings,
-    disabled_ai_monitoring_settings,
+from testing_support.ml_testing_utils import (
+    add_token_count_to_events,
     events_sans_content,
     llm_token_count_callback,
+    disabled_ai_monitoring_record_content_settings,
+    disabled_ai_monitoring_settings,
+    set_trace_info,
 )
 from testing_support.fixtures import (
     reset_core_stats_engine,
@@ -104,7 +105,7 @@ def test_openai_embedding_sync_no_content(set_trace_info, sync_openai_client):
 
 @reset_core_stats_engine()
 @override_llm_token_callback_settings(llm_token_count_callback)
-@validate_custom_events(add_token_count_to_event(embedding_recorded_events))
+@validate_custom_events(add_token_count_to_events(embedding_recorded_events))
 @validate_custom_event_count(count=1)
 @validate_transaction_metrics(
     name="test_embeddings_v1:test_openai_embedding_sync_with_token_count",
@@ -183,7 +184,7 @@ def test_openai_embedding_async_no_content(loop, set_trace_info, async_openai_cl
 
 @reset_core_stats_engine()
 @override_llm_token_callback_settings(llm_token_count_callback)
-@validate_custom_events(add_token_count_to_event(embedding_recorded_events))
+@validate_custom_events(add_token_count_to_events(embedding_recorded_events))
 @validate_custom_event_count(count=1)
 @validate_transaction_metrics(
     name="test_embeddings_v1:test_openai_embedding_sync_with_token_count_async",
